requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.2
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
ffmpy==0.6.1
filelock==3.18.0
fsspec==2025.7.0
google-auth==2.40.3
google-genai==1.29.0
gradio==5.42.0
gradio-client==1.11.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
ipython==8.12.3
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.2
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==6.31.1
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
ruff==0.12.8
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
tenacity==9.1.2
tinycss2==1.4.0
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
yarg==0.1.9
